Revert "Move launch parameter sanitation back from separate files into anonymous namespace"
This reverts commit 88502390d9d46751edf451602e0372b897a25674.
