+++ libvast/src/system/configuration.cpp
@@ -23,6 +23,7 @@
 #include "vast/detail/stable_set.hpp"
 #include "vast/detail/string.hpp"
 #include "vast/detail/system.hpp"
+#include "vast/launch_parameter_sanitation.hpp"
 #include "vast/logger.hpp"
 #include "vast/plugin.hpp"
 #include "vast/synopsis_factory.hpp"
@@ -285,42 +286,6 @@ caf::expected<caf::settings> to_settings(record config) {
   return to<caf::settings>(config);
 }
 
-auto generate_default_value_for_argument_type(std::string_view type_name) {
-  if (type_name.starts_with("uint") || type_name.starts_with("int")
-      || type_name.starts_with("long")) {
-    return "0";
-  } else if (type_name == "timespan") {
-    return "0s";
-  } else if (type_name.starts_with("list")) {
-    return "[]";
-  }
-  VAST_ASSERT(false && "option has type with no default value support");
-  return "";
-}
-
-void sanitize_long_form_argument(std::string& argument,
-                                 const vast::command& cmd) {
-  auto dummy_options = caf::settings{};
-  auto [state, _] = cmd.options.parse(dummy_options, {argument});
-  if (state == caf::pec::not_an_option) {
-    for (const auto& child_cmd : cmd.children) {
-      sanitize_long_form_argument(argument, *child_cmd);
-    }
-  } else if (state == caf::pec::missing_argument) {
-    auto name = argument.substr(2, argument.length() - 3);
-    auto option = cmd.options.cli_long_name_lookup(name);
-    if (!option) {
-      // something is wrong with the long name options:
-      // reveal this during the actual parsing.
-      return;
-    }
-    auto option_type = option->type_name();
-    auto options_type_default_val
-      = generate_default_value_for_argument_type(option_type.data());
-    argument.append(options_type_default_val);
-  }
-}
-
 } // namespace
 
 std::vector<std::filesystem::path>
@@ -396,7 +361,7 @@ caf::error configuration::parse(int argc, char** argv, const command& root) {
   command_line.assign(argv + 1, argv + argc);
   for (auto& argument : command_line) {
     if (argument.starts_with("--")) {
-      sanitize_long_form_argument(argument, root);
+      launch_parameter_sanitation::sanitize_long_form_argument(argument, root);
     }
   }
   // Translate -qqq to -vvv to the corresponding log levels. Note that the lhs
@@ -480,7 +445,7 @@ caf::error configuration::parse(int argc, char** argv, const command& root) {
   // Newly added plugin arguments from environment variables
   // may contain empty values - sanitize them.
   for (auto& plugin_arg : plugin_args) {
-    sanitize_long_form_argument(plugin_arg, root);
+    launch_parameter_sanitation::sanitize_long_form_argument(plugin_arg, root);
   }
   // Copy over the specific plugin options.
   std::move(plugin_opt, command_line.end(), std::back_inserter(plugin_args));
